refactor: onMapNameAssigned to onImpactProducerAssigned
