+++ server/interfaces/messages/ImpactProducerAssigned.ts
@@ -1,6 +1,6 @@
 export type MapFeatureKinds = 'raw_capture' | 'capture'
 
-export default interface MapNameMessage {
+export default interface ImpactProducerAssigned {
   type: string
   impact_producer_id: string
   map_feature_ids: string[]
