+++ server/messaging/registerEventHandlers.ts
@@ -1,5 +1,5 @@
 import onCaptureCreated from 'events/onCaptureCreated'
-import onMapNameAssigned from 'events/onMapNameAssigned'
+import onImpactProducerAssigned from 'events/onImpactProducerAssigned'
 import onRawCaptureCreated from 'events/onRawCaptureCreated'
 import onTokenAssigned from 'events/onTokenAssigned'
 import { subscribe } from 'messaging/broker'
@@ -12,7 +12,7 @@ export default async function registerEventHandlers() {
     await subscribe(SubscriptionNames.TOKEN_ASSIGNED, onTokenAssigned)
     await subscribe(
       SubscriptionNames.IMPACT_PRODUCER_ASSIGNED,
-      onMapNameAssigned,
+      onImpactProducerAssigned,
     )
   } catch (e) {
     console.error('Get error when handling message:', e)
